chore: small refactoring
